requirements.txt
streamlit==1.29.0
tensorflow==2.13.0
keras==2.13.1
numpy==1.24.3
pillow==10.0.1
opencv-python-headless==4.8.1.78
scikit-learn==1.3.2
matplotlib==3.8.0
